asda

+++ Routes/Notes.route.js
@@ -7,39 +7,45 @@ const jwt = require("jsonwebtoken");
 notesrouter.get("/", async (req, res) => {
   try {
     const notes = await NotesModel.find();
-    res.status(200).send(notes);
-  } catch (error) {}
+    res.send(notes);
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 notesrouter.post("/addnotes", async (req, res) => {
   try {
-    await jwt.verify(req.headers.authorization, "ganesh", (err, decoded) => {
-      if (decoded) {
-        let user = new NotesModel(req.body);
-        user.save();
-        res.send("notes added success");
-      } else {
-        res.send("login required");
-      }
-    });
+    let user = new NotesModel(req.body);
+    await user.save();
+    res.send("notes added success");
   } catch (err) {
     console.log(err);
   }
 });
 
 notesrouter.patch("/updatenotes/:id", async (req, res) => {
+  const user = await NotesModel.findOne({ _id: req.params.id });
   try {
-    const user = await NotesModel.findByIdAndUpdate(req.params.id, req.body);
-    res.send("notes updated success");
+    if (user.userId != req.body.userId) {
+      res.send("user not Authorised");
+    } else {
+      await NotesModel.findByIdAndUpdate(req.params.id, req.body);
+      res.send("notes updated success");
+    }
   } catch (error) {
     console.log(error);
   }
 });
 
 notesrouter.delete("/deletenotes/:id", async (req, res) => {
+  const user = await NotesModel.findOne({ _id: req.params.id });
   try {
-    const user = await NotesModel.findByIdAndDelete(req.params.id);
-    res.send("notes deleted success");
+    if (user.userId != req.body.userId) {
+      res.send("user not Authorised");
+    } else {
+      const user = await NotesModel.findByIdAndDelete(req.params.id);
+      res.send("deleted success");
+    }
   } catch (error) {
     console.log(error);
   }
